refactor folder structure with new kedro "framework" folder
