+++ kedro_mlflow/framework/hooks/pipeline_hook.py
@@ -4,12 +4,12 @@ from typing import Any, Dict, Union
 
 import mlflow
 import yaml
-from kedro.hooks import hook_impl
+from kedro.framework.hooks import hook_impl
 from kedro.io import DataCatalog
 from kedro.pipeline import Pipeline
 from kedro.versioning.journal import _git_sha
 
-from kedro_mlflow.context import get_mlflow_conf
+from kedro_mlflow.framework.context import get_mlflow_conf
 from kedro_mlflow.mlflow import KedroPipelineModel
 from kedro_mlflow.pipeline import PipelineML
 from kedro_mlflow.utils import _parse_requirements
