+++ kedro_mlflow/template/project/run.py
@@ -30,10 +30,10 @@
 from pathlib import Path
 from typing import Dict
 
-from kedro.context import KedroContext, load_context
+from kedro.framework.context import KedroContext, load_context
 from kedro.pipeline import Pipeline
 
-from kedro_mlflow.hooks import MlflowNodeHook, MlflowPipelineHook
+from kedro_mlflow.framework.hooks import MlflowNodeHook, MlflowPipelineHook
 from {{cookiecutter.python_package}}.pipeline import create_pipelines
